+++ shared/crypto/sub-nav/index.native.tsx
@@ -1,0 +1,10 @@
+import * as React from 'react'
+import * as KB from '../../common-adapters'
+
+const SubNav = (props: Props) => {
+  return (
+
+  )
+}
+
+export default SubNav
